feat: 实现礼物编辑；添加文件上传记录，文件大小限制5MB；

+++ src/main/java/com/xiaoyi/birthdayreminder/service/impl/GiftServiceImpl.java
@@ -2,13 +2,14 @@ package com.xiaoyi.birthdayreminder.service.impl;
 
 import com.xiaoyi.birthdayreminder.context.BaseContext;
 import com.xiaoyi.birthdayreminder.mapper.GiftMapper;
-import com.xiaoyi.birthdayreminder.pojo.Result;
 import com.xiaoyi.birthdayreminder.pojo.dto.GiftDTO;
 import com.xiaoyi.birthdayreminder.pojo.dto.GiftItemDTO;
 import com.xiaoyi.birthdayreminder.pojo.dto.GiftLikeDTO;
 import com.xiaoyi.birthdayreminder.pojo.entity.Gift;
 import com.xiaoyi.birthdayreminder.pojo.entity.PageBean;
+import com.xiaoyi.birthdayreminder.service.FileService;
 import com.xiaoyi.birthdayreminder.service.GiftService;
+import com.xiaoyi.birthdayreminder.utils.AliOssUtil;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -23,6 +24,12 @@ public class GiftServiceImpl implements GiftService {
     @Autowired
     private GiftMapper giftMapper;
 
+    @Autowired
+    private AliOssUtil aliOssUtil;
+
+    @Autowired
+    private FileService fileService;
+
     @Override
     public void add(GiftDTO giftDTO) {
         Gift gift = new Gift();
@@ -31,18 +38,15 @@ public class GiftServiceImpl implements GiftService {
         gift.setUpdateTime(LocalDateTime.now());
         String creator = BaseContext.getCurrentId();
         gift.setCreator(creator);
-        gift.setStarCount(0);
-        gift.setLikeCount(0);
 
         giftMapper.insert(gift);
-
     }
 
     @Override
     public PageBean list(Integer start,Integer pageSize,boolean isMine,boolean isCollect) {
         String creatorId = BaseContext.getCurrentId();
-        List<GiftItemDTO> giftList =null;
-        Long count = null;
+        List<GiftItemDTO> giftList;
+        Long count;
 
         if(isCollect){
             giftList = giftMapper.listByCollect(start, pageSize, creatorId);
@@ -87,11 +91,37 @@ public class GiftServiceImpl implements GiftService {
 
     @Override
     public GiftItemDTO detail(Integer id) {
-        return giftMapper.detail(id);
+        GiftItemDTO giftItemDTO =  giftMapper.detail(id);
+        Boolean isMine = giftItemDTO.getCreator().equals(BaseContext.getCurrentId());
+        giftItemDTO.setIsMine(isMine);
+        return giftItemDTO;
     }
 
     @Override
     public void delete(Integer id) {
-        giftMapper.delete(id);
+        GiftItemDTO giftItemDTO = this.detail(id);
+        if(giftItemDTO.getCreator().equals(BaseContext.getCurrentId())){
+            String image=giftItemDTO.getImage();
+            String objectName=  aliOssUtil.getObjectName(image);
+            aliOssUtil.delete(objectName);
+            fileService.deleteByUrl(image);
+
+            giftMapper.delete(id);
+        }
+    }
+
+    @Override
+    public void update(GiftDTO giftDTO) {
+        GiftItemDTO giftItemDTO = this.detail(giftDTO.getId());
+        Gift gift = new Gift();
+        gift.setId(giftItemDTO.getId());
+        gift.setName(giftDTO.getName());
+        gift.setDescription(giftDTO.getDescription());
+        gift.setImage(giftDTO.getImage());
+
+        if(giftItemDTO.getCreator().equals(BaseContext.getCurrentId())){
+            gift.setUpdateTime(LocalDateTime.now());
+            giftMapper.update(gift);
+        }
     }
 }
